내모임 화면 다가오는 일정 api 연동

+++ core/network/src/main/java/com/onmoim/core/network/model/meeting/MeetingDto.kt
@@ -23,9 +23,9 @@ data class MeetingDto(
     @SerialName("joinCount")
     val joinCount: Int,
     @SerialName("latitude")
-    val latitude: Int,
+    val latitude: Double,
     @SerialName("longitude")
-    val longitude: Int,
+    val longitude: Double,
     @SerialName("placeName")
     val placeName: String,
     @SerialName("startAt")
